Input 분리
